Use UTC when converting to QDateTime for graph axis
QDateTime::fromMSecsSinceEpoch returns a QDateTime in the local time zone, but the ms input was created in UTC.  This causes the day part to sometimes display off by a day on the graph x axis.  This appears to be caused by a breaking change in QT 5.  Previously, fromMSecsSinceEpoch returned a QDateTime in UTC, but in QT 5, it returns the date time in the local time zone.