Update all calls to fromMSecsSinceEpoch to specify UTC
QT 5 changed the behavior of QDateTime::fromMSecsSinceEpoch to return the result in local time zone, whereas in QT 4, it returned UTC.  On systems that do not support time zones, the QT 5 version still returns UTC.  But for all other systems, the api change causes confusion in the SH UI because some date and time values are displayed in UTC instead of the local time zone.  This manifests itself when the user is in USA which has a negative UTC offset.  When selecting a date range to display in the overview screen, the displayed dates appear to be one day behind the selected date range.

For consistently, SH should always use UTC internally and only convert to the local time zone when displaying data to the user.  This will ensure that the time zone information is preserved correctly when the UTC offset of the user's machine changes due to DST changes or traveling.

There are a few calls to fromMSecsSinceEpoch which should be using local time, and those will be updated in future commits.